Fix ability to cross-compile extension modules. (#14)
This solves for some very confusing configuration changes in
`rules_rust.cargo_build_script`. Previously cross-compilation would not
work despite having the ability to always generate a
`pyo3_build_config.txt` for the right configuration. In practice, it's
only ever generated for the `exec` configuration. Because it's compiled
into `pyo3-build-config`, which is then used in other build scripts, the
built config will always e in the `exec` configuration because build
script scripts are always built for it (see
[@rules_rust//cargo/private/cargo_build_script_wrapper.bzl](https://github.com/bazelbuild/rules_rust/blob/bad49c4a1139afdb257fed0af6bd1dc20a6c21a1/cargo/private/cargo_build_script_wrapper.bzl#L160-L210)
and
[@pyo3//pyo3-build-config/src/lib.rs](https://github.com/PyO3/pyo3/blob/672efdc3ed19ea575b915daa33333de664331f47/pyo3-build-config/src/lib.rs#L106-L131)).
In order to ensure the correct settings are used in the target
configuration they must be provided to each build script that needs it.
Thus you'll see that new annotations were added to `pyo3-ffi`, which
previously was trying to use config data inlined into
`pyo3-build-config`.