Fix abi3 using python headers (#42)
Even though abi3, abi3-py311, abi3-py312 and extension-package are used
as dependencies, the compiled .so (.dylib) file is not abi3 compatible
(the same .so file can't be run on python 3.12, 3.13).

By running `bazel 'build' '//pyo3/private/tests/string_sum:string_sum'`
and generating `string_sum.so` we can see a link to
`@rpath/libpython3.11.dylib (compatibility version 3.11.0, current
version 3.11.0)` using otool on mac (ldd on linux):

```
otool -L string_sum.so
string_sum.so:
        bazel-out/darwin_arm64-opt/bin/pyo3/private/tests/string_sum/libstring_sum.dylib (compatibility version 0.0.0, current version 0.0.0)
        @rpath/libpython3.11.dylib (compatibility version 3.11.0, current version 3.11.0)
        /usr/lib/libiconv.2.dylib (compatibility version 7.0.0, current version 7.0.0)
        /usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 1351.0.0)
        /usr/lib/libc++.1.dylib (compatibility version 1.0.0, current version 1900.178.0)
```

This is due to the usage of
`@rules_python//python/cc:current_py_cc_libs`. It can be replaced with
`@rules_python//python/cc:current_py_cc_headers` and removing the link
to `@rpath/libpython3.11.dylib`.

With the changes in this PR, the otool shows:

```
otool -L string_sum.so                                                                  
string_sum.so:
        bazel-out/darwin_arm64-opt/bin/pyo3/private/tests/string_sum/libstring_sum.dylib (compatibility version 0.0.0, current version 0.0.0)
        /usr/lib/libiconv.2.dylib (compatibility version 7.0.0, current version 7.0.0)
        /usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 1351.0.0)
        /usr/lib/libc++.1.dylib (compatibility version 1.0.0, current version 1900.178.0)
```

rustc flags had to be added to work on MacOS. I would be happy to
refactor this to the feature flag so that abi3 would not be used
automatically if not needed.